requirements.txt
anyio==3.6.2
app==0.0.1
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
cryptography==39.0.2
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.92.0
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multipart==0.2.4
mysql-connector-python==8.0.32
orjson==3.8.10
passlib==1.7.4
protobuf==3.20.3
pycparser==2.21
pydantic==1.10.5
PyJWT==2.6.0
PyMySQL==1.0.2
pyodbc==4.0.35
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
